--- a/index-matching.ts
+++ b/index-matching.ts
@@ -1,6 +1,6 @@
 export { default as ReverseIndex } from './src/engine/reverse-index';
-export {default as matchNetworkFilter } from './src/matching/network';
+export { default as matchNetworkFilter } from './src/matching/network';
 export { parseNetworkFilter } from './src/parsing/network-filter';
-export { mkRequest } from './src/request/interface';
+export { default as Request } from './src/request';
 export { tokenize, createFuzzySignature } from './src/utils';
 export { compactTokens, hasEmptyIntersection, mergeCompactSets } from './src/compact-set';