--- a/src/engine/engine.ts
+++ b/src/engine/engine.ts
@@ -262,6 +262,20 @@ export default class FilterEngine {
     );
   }
 
+  public matchAll(rawRequest: Partial<IRequestInitialization>): Set<NetworkFilter> {
+    const request = new Request(rawRequest);
+
+    const filters = [];
+    if (request.isSupported) {
+      filters.push(...this.importants.matchAll(request));
+      filters.push(...this.filters.matchAll(request));
+      filters.push(...this.exceptions.matchAll(request));
+      filters.push(...this.redirects.matchAll(request));
+    }
+
+    return new Set(filters);
+  }
+
   public match(
     rawRequest: Partial<IRequestInitialization>,
   ): {