… CHANGELOG.md
… assets/easylist-downloads.adblockplus.org/antiadblockfilters.txt
… assets/easylist.to/easylist/easylist.txt
… assets/easylist.to/easylist/easyprivacy.txt
… assets/easylist.to/easylistgermany/easylistgermany.txt
… assets/pgl.yoyo.org/adservers/serverlist.txt
… assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/badware.txt
… assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/filters.txt
… assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/resource-abuse.txt
… assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/unbreak.txt
… bench/dataset/collect_metrics.js
… bench/dataset/create_list.js
… bench/macro.js
… bench/micro.js
… bench/utils.js
… dist/adblocker.es.js
… dist/adblocker.umd.js
… dist/index-matching.d.ts
… dist/index.d.ts
… dist/pattern-matching.es
… dist/src/dynamic-data-view.d.ts
… dist/src/engine/bucket/network.d.ts
… dist/src/engine/engine.d.ts
… dist/src/engine/reverse-index.d.ts
… dist/src/matching/network.d.ts
… dist/src/parsing/cosmetic-filter.d.ts
… dist/src/parsing/interface.d.ts
… dist/src/parsing/network-filter.d.ts
… dist/src/request.d.ts
… dist/src/request/interface.d.ts