This PR is a work-in-progress aiming at speeding up the matching of content-blocker. It is the result of collecting and analyzing metrics from running an instrumented version of the adblocker against all network requests seen on the top 50k domains.

Some performance metrics:

**Index**:
* Reverse index had 28709 buckets
* 26611 of buckets have only one filter inside (92%)
* 1490 of buckets have two filters (5%)

**Requests**:
* For blocked requests, we only needed to inspect 3 filters on average
* For allowed requests, we had to inspect 7.7 filters on average
* Processing a request took 0.019 ms on average (0.016 ms when no match)

**Filters**:
* 11485 filters were inspected at least once (36%)
* 4336 filters matched a request at least once (13%)

**i7 U6600**

- Easylist (**x3 speedup**)
  * Min match: 0.004 => 0.003ms
  * Max match: 1.4 => 1.0ms
  * Avg match: 0.018 => 0.012ms

  * Min no match: 0.001732 => 0.00089ms
  * Max no match: 313.360989 =>**1.9ms**
  * Avg no match: 0.03 => **0.011ms**
- Offers (**x10 speedup**)
  * Min no match: 0.07 => **0.0006ms**
  * Max no match: **8.5** => 11ms
  * Avg no match: 0.08 => **0.008ms**
- Mobile (**x2 speedup**)
  * Min match: **0.001** => 0.002ms
  * Max match: 1.18 => **1.07ms**
  * Avg match: 0.010 => 0.010ms

  * Min no match: 0.0011 => **0.008ms**
  * Max no match: 81.0=> **1.5ms**
  * Avg no match: 0.014 => **0.008ms**

On a `nanopi` Neo2 (CPU ARM Cortex-A7 1.2Ghz):
- Easylist (x2 on average, but massive reduction on outliers if no match)
  * Min match: 0.081206 => 0.05ms
  * Max match: 13.949639 => 7ms
  * Avg match: 0.22 => **0.17ms**

  * Min no match: 0.034708 => **0.01ms**
  * Max no match: 1536 => **7ms**
  * Avg no match: 0.30 => **0.14ms**
- Mobile (Mostly the same except when there is no match)
  * Min match: **0.028416** => 0.05ms
  * Max match: **11.863192** => 29ms
  * Avg match: **0.14** => 0.15ms

  * Min no match: 0.028666 => **0.016ms**
  * Max no match: 432.85874 => **15ms**
  * Avg no match: 0.18 => **0.12ms**
- Offers (**x20** faster on average)
  * Min no match: 1.85667 => **0.015ms**
  * Max no match: 96.855703 => **28ms**
  * Avg no match: 2.06 => **0.11ms**

**CHANGELOG**:

- [BREAKING] format of serialized engine has been changed to store less data
- [BREAKING] `id` attribute from filters has been remove, use `getId()` instead (please note that the `id` is not stored internally anymore, but generated every time `getId()` is called).
```js
// Bad
filter.id
// Good
filter.getId()
```
- [BREAKING] values returned by `getId()` will differ from values stored in
  the `id` attribute for identical filters (the algorithm is now different
  and will do less work).
- [BREAKING] domains specified in `$domains=` option are now stored hashed
  instead of as string, and can only be retried in their original form if
  `debug` flag is used in `FiltersEngine`
- [BREAKING] `fastTokenizer` will now only consider tokens longer than 1
- [BREAKING] `fastTokenizer` will now only tokenize up to 2048 characters from URLs
- [BREAKING] hashes produced by `fastHash` and `fastHashBetween` will not
  match what was produced by the same function before this change (the seed
  and hashing algorithm was slightly changed for speed).
- [BREAKING] un-initialized attributes of filters instances
  (`CosmeticFilter` and `NetworkFilter`) will have value `undefined`
  instead of `null` or empty string like before. It is recommended
  to use accessors (e.g.: `filter.getHostname()` instead of
  `filter.hostname`) to access internal attributes, as they will
  always return consistent types and fall-back to meaningful defaults.
```js
// Bad
filter.redirect
filter.filter
filter.hostname

// Good
filter.getRedirect()
filter.getFilter()
filter.getHostname()
```
- [BREAKING] a new `Request` abstraction supersedes `IRequest` and
  `IRawRequest`. This new class offers a more consistent experience to work
  will requests.
```js
new Request({ url })
new Request({ url, sourceUrl })
new Request({ url, sourceUrl, type: 'string' })
new Request({ url, hostname, domain, type: 'string' })
```
- [BREAKING] remove support for `hosts` format (e.g.: `127.0.0.1 domain`),
  since servers blocklists can also be exported in hostname anchored format
  (e.g.: `||domain^$third-party`). This simplifies the parsing logic.
- [BREAKING] remove the following unused legacy request types:
    * `fromFetch`
    * `fromDTD`
    * `fromXLST`
    * `fromBeacon`
    * `fromCSP`
- [BREAKING] `cpt` (Content Policy Type of requests) is now called `type`,
  to match the terminology of the WebRequestAPI.
```js
// Bad
request.cpt
new Request({ cpt })

// Good
request.type
new Request({ type })
```
- Requests can now use `type` as a string or number (e.g.: `script` or `2`).
```js
// Both are equivalent
new Request({ type: 2, url, sourceUrl })
new Request({ type: 'script', url, sourceUrl })
```
- Optimized and simplified implementation of `parseJSResource` (~4 times faster)
- Optimized matching of some kinds of filters to prevent any string copy (reduced the number of calls to `slice`, `substr` and `substring`)
- Optimized buckets ordering by moving matching filters towards the
  beginning of the array. This results in generic filter being tried first.
- Optimized some classes of filters sharing the same pattern and options,
  with different domains. They are now fused into a single filter. For
  example, the following filters:
    * `|https://$script,domain=downloadpirate.com`
    * `|https://$script,domain=dwindly.io`
    * `|https://$script,domain=intoupload.net`
    * `|https://$script,domain=linkshrink.net`
    * `|https://$script,domain=movpod.in`
    * `|https://$script,domain=povw1deo.com|povwideo.net|powvideo.net`
    * `|https://$script,domain=sendit.cloud`
    * `|https://$script,domain=sfiles.org|suprafiles.me|suprafiles.net|suprafiles.org`
    * `|https://$script,domain=streamplay.to`
    * `|https://$script,domain=userscloud.com`
    * `|https://$script,domain=yourporn.sexy`
will be optimized into: `|https://$script,domain=dwindly.io|movpod.in|...|yourporn.sexy`
- `tokenize` will now allow `%` as part of tokens for filters
- `CosmeticFilter` now support the new `+js()` syntax to inject scripts
- `NetworkFilter`'s `getTokens()` method will now return more tokens in some
  cases. For example, if only one domain is specified in the `$domain=`
  option, then it can be used as a token (before we would only use the pattern part of each filter to extract tokens).
- In case a `NetworkFilter` has no token available (e.g.:
  `$image,domain=ads.com`), then it can be indexed using the domains
  specified in the `$domain=` option, if any.
- Filters of the form `*pattern` (regex) are now optimized into `pattern` (plain)
- Filters of the form `|http://` or `|https://` or `|http*://` are now
  optimized using the newly introduced `http` and `https` options. The
  `Request` instances will now say if they are `http` or `https`, and this
  saves string comparisons while matching.
- Fixed a bug where javascript resources were serialized twice
- Serialization can now be performed even after `engine` has been optimized
- Addition of a `serialize` method on `FiltersEngine` class